Fixed some basic compiler warnings in the x-plane plugin code so it should compile cleanly now. also added comments and tried to simplify the code involving the acceleration and rate conversions, so hopefully other people can understand it and help me get it working properly.
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@262 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ Tools/XP_UDB_HILSIM/stdafx.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #define WIN32_LEAN_AND_MEAN		// Exclude rarely-used stuff from Windows headers
-#define PI 3.141592653589793238462643
+#define PI (float)3.141592653589793238462643
 #define NUMLOOPS 6
 #define R_EARTH 6371000
